Use package.json browser field for browser-friendly builds
