+++ js/src/Ice/FileLogger.browser.js
@@ -1,0 +1,5 @@
+// Copyright (c) ZeroC, Inc.
+
+// Mockup FileLogger for browser environment
+let FileLogger = null;
+export { FileLogger };
