+++ js/src/Ice/TcpTransceiver.browser.js
@@ -1,0 +1,5 @@
+// Copyright (c) ZeroC, Inc.
+
+// Mockup TcpTransceiver for browser environment
+let TcpTransceiver = null;
+export { TcpTransceiver };
