--- js/test/Common/fs.mockup.js
@@ -1,8 +1,0 @@
-// A mockup for NodeJS fs module
-
-const readFileSync = {};
-const fs = {
-    readFileSync: readFileSync,
-};
-export default fs;
-export { readFileSync };
