--- js/test/Common/net.mockup.js
@@ -1,4 +1,0 @@
-// A mockup for NodeJS net module
-
-const net = {};
-export default net;
