--- js/test/Common/path.mockup.js
@@ -1,9 +1,0 @@
-// A mockup for NodeJS path module
-
-const join = {};
-const path = {
-    join: join,
-};
-
-export { join };
-export default path;
